# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DR TODAY API Documentation",
        "description": "API documentation for DR TODAY application",
        "contact": {
            "email": "admin@dr-today.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "DR TODAY API Server"
        }
    ],
    "paths": {
...
        "/api/companies": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Get all Companies",
                "operationId": "ac82cf1c054ef7cbe0c010c867c5b527",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            },
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "Create a new Companies",
                "operationId": "cb75885c6c65fc26c387eaed207996c5",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/companies/{company}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Get all {company}",
                "operationId": "747b19097c1b918066cbdd7ff4e1e77e",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            },
            "put": {
                "tags": [
                    "Company"
                ],
                "summary": "Update a {company} by ID",
                "operationId": "d41e335de1a0343da7845bd470f80c5e",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/get_company/{id}": {
            "get": {
                "tags": [
                    "Get_company"
                ],
                "summary": "Get a Get_company by ID",
                "operationId": "9c50388e9a2cc058252b848e867ab408",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/companies/delete/{id}": {
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "Create a new Delete",
                "operationId": "d414529ae0b1924bab1cf25a5ab7254d",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/companies/restore/{id}": {
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "Create a new Restore",
                "operationId": "bea8451d2685b8e504e9b306618436d5",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}